requirements.txt
# conda create --prefix=./.venv python=3.12

tensorflow
pandas
scikit-learn
matplotlib
seaborn
statsmodels
tensorflow-datasets
jupyter
pandas-datareader
yfinance
alpha_vantage
cufflinks
nltk
spacy[transformers]
imblearn
datetime
fastapi
uvicorn
motor
passlib
pydantic
email-validator
bcrypt
python-jose[cryptography]
python-multipart
python-dotenv
httpx
PyPDF2
langchain
langchain-community
langchain-openai
openai
streamlit
wikipedia
numexpr
langgraph
google-search-results
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
langchain-google-genai
tf-keras
InstructorEmbedding
sentence-transformers
langchain-huggingface
